reorganize frontend modules
