+++ frontend/src/modules/dashboard/pages/DashboardPage.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { getOrdenesTrabajo } from '../../services/api';
+import { getOrdenesTrabajo } from '../../../services/api';
 import { 
   ClockIcon, 
   CheckCircleIcon,
