+++ frontend/src/modules/landing/components/Footer.jsx
