+++ frontend/src/modules/landing/components/Hero.jsx
