+++ frontend/src/modules/landing/pages/LandingPage.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 
-import Header from './components/Header';
-import Hero from './components/Hero';
-import Servicios from './components/Servicios/Servicios';
-import SobreMi from './components/SobreMi';
-import Contacto from './components/Contacto';
-import Footer from './components/Footer';
+import Header from '../components/Header';
+import Hero from '../components/Hero';
+import Servicios from '../components/Servicios/Servicios';
+import SobreMi from '../components/SobreMi';
+import Contacto from '../components/Contacto';
+import Footer from '../components/Footer';
 
 function LandingPage() {
   return (
